backend/requirements.txt
fastapi>=0.109.0
uvicorn==0.34.0
pydantic>=2.7.2
python-dotenv==1.0.0
beautifulsoup4==4.12.3
lxml==5.1.0
aiohttp>=3.9.0
aiohttp-socks>=0.8.3
httpx>=0.28.1
websockets>=13.0.0,<15.1.0
h11<0.15,>=0.13
